Issue #83: Include and use the Source Code pro font for the editor and REPL window.
I've included some variants of Source Code Pro (It's SIL licensed so no issues there)
and made the interface load and use that.  Because of some changes in the way the
fonts are specified, I took advantage of the existing Font class to include
some of the loading logic.

Tests updated etc to match