Use Python regex to highlight matches
The existing code used Scintilla's built-in findFirst method which changes the active selection as it goes. The attempt to replace the original selection was causing issues when selecting backwards with the keyboard. While it would probably be possible to get the right logic in place by getting the right combination of Scintilla's anchor and caret position, I've simply bypassed the problem by using Python's regex .finditer function to identify matching spans and converting those to Scintilla ranges.