Support for custom MicroPython runtimes
By default, Mu uses the MicroPython runtime that is bundled in uFlash.
This change enables custom runtimes to be used. The user can specify
the runtime in the settings.json file, e.g.:
"runtime_hex": "custom_runtime.hex"
The file can be a fully-qualified path and file name, or it can
just be a file name in which case it should be located in the
workspace directory.
This change also introduces version 1.0.8 of uFlash