CI Linux: Enable tests on the Linux Travis builds
There was a problem runing the tests on the Linux builds on Travis.
This has been a problem for quite a long time, and in different
points it yielded different, unuseful errors.

Turns out it was Qt failing to launch when there isn't a frame
buffer available, so using xvfb on Linux.

We've also moved from setting up our own Python environment to use
the ones provided by Travis itself, so we are testing on Python 3.5
and 3.6.