Make sure that test failure modes happen correctly on Windows
To make Mu more robust, the new Editor._abspath functionality will log and move on if it fails to process in some way the path being passed to it. It's difficult to test this on Windows because pretty much anything (including None) will succeed when passed to os.abspath. Since we're catching all Exception, just in case, the easiest thing is to force a Type Error.

One of the test cases was fixed by 2c5b70abfc301df1c9f356a5be962e5a0b939dae but used a lambda expression to force failure. I've switched this out with a simple integer, since I had to  look twice to see if the lambda was doing something subtle (it wasn't), whereas the integer is obviously an invalid filename.