Detect an invalid codec in an encoding cookie
When Mu saves a file it detects the presence of an encoding cookie on the first line. If present, the codec specified in the cookie is used. If that is an invalid codec (possibly the result of an autosave while the cookie is still being written) then log a warning and drop back to the Mu default codec.