Workaround for #552: Fix black text issue with qtscintilla
On mojave, using a bundled (non-system-installed) font with
qtscintilla (qt5) results in text being black regardless of
the theme color.  This is not ideal with dark themes where
the text becomes invisible.

To work around the problem, detect Mojave system (any MacOS version 18)
and change the default font to Monaco (a system monospace font)

We still extract and load the Source Code Pro font into the QT
font database, to keep the code changes to a minimum, but the editor
uses Monaco, which allows syntax highlihting to work