Automatically close an open repl/plotter pane when serial port disapears
While polling for available USB devices, check to see if ones that
disappear happen to be in use by a repl or plotter pane. If so, close
those panes to let the user know that they are no longer connected and
close the serial port so that Qt doesn't constantly poll the file
descriptor.

I'm not sure this is the right plan, but if we don't do anything when
the device disappears, then the user will have to close and re-open
the REPL when the device is reconnected, which can be confusing. In
addition, Qt will spin on the missing serial device.

Signed-off-by: Keith Packard <keithp@keithp.com>

---

v2:
	Add tests for logic and interface/main to make sure devices
	which disappear will be tracked correctly