Set minimum version for virtualenv dependency.
Older versions of virtualenv create a pip file (a python file
without extension) in the venv bin folder that uses a "normal"
shebang pointing to the python executable inside the venv.
If the path contains spaces the shell shebang parsing can break
and stop processing it after the first space (even if the space
is scaped with `\ `). When this happens it means that mu.virtualenv
fails trying to call venv/bin/pip with an error like:
```
bad interpreter: No such file or directory
```

From version 16.* of virtualenv the shebang has been changed to
use a format that should work fine with spaces:

```
'''exec' "/path/to/venv with spaces/bin/python3" "$0" "$@"
' '''
```