venv: Install each venv wheel individually.
This also provides more granular logs and decreases the chance of
having a timeout of the pip install process. In the previous
implementation the 3 minutes configure for the timeout might not
have been long enough, in this case we have 3 min per wheel which
should be enough for individal install.