uFlash: Ensure the fs hex string injected into the Universal Hex is memory aligned.
This is a manual cherry-pick of the uFlash PR 74:
https://github.com/ntoll/uflash/pull/74/

From the commit:

USB uses 512 byte blocks (or a multiple of 512), and DAPLink depends
on some Universal Hex records (like the Block Start record) to be
well aligned within their "section" (the part that contains micro:bit
V1 or V2 code).

This commit ensures the generated filesystem hex records follow the
same alignment, so that when they are injected into the Universal Hex
this is maintained.

This should have been done as part of the original Universal Hex
implementation, as it is part of the official spec, but I missed it:
https://tech.microbit.org/software/spec-universal-hex/

uFlash itself *does not* ensure the Universal Hex is aligned, but it
will be if it was genefrated by the microbit-universal-hex library:
https://github.com/microbit-foundation/microbit-universal-hex

As uFlash does generated aligned hex strings it preserves whatever
alignment the original Universal Hex sting has.