Set ipykernel version to be < 6.0 for macOS 10.13 compat.
This is a qtconsole dependency, which only declares the min version,
so when it gets the latest version available during packaging it
downloads v6, which adds a dependency that is not compatible with
macOS 10.13 High Sierra.
This dependency, `debugpy`, is only needed for the debugging
feature added to iPython in v6, so it's fine not to have it.

After the Mu v1.1 release we'll have to update PyQt5 to a newer
version that will drop support for macOS 10.13, so at that point
we can unpin `ipykernel`.