Allow load_pixmap to be given a target size
Many pixmap users end up rescaling the loaded image to suit the
UI. This negates the advantages of SVG icons as a temporary image is
generated and then scaling happens from that.

Avoid this by allowing pixmap consumers to specify the target size,
and if specified, use the icon loader and then scaling from there
instead of just using the pixmap loader.

Signed-off-by: Keith Packard <keithp@keithp.com>