On uPy files error close panel before showing error.
Fixes https://github.com/mu-editor/mu/issues/2262.

This issue happens because multiple errors while connecting with
a MicroPython device can try to close/destroy de files panel.
So there is a bit of a race condition and the second error might
try to close an already destroyed files panel.

The way to replicate this issue is to find a way to get an error
while reading the files from a MicroPython device, and then to cause
another connection error by, for example, unplugging the device.
The simplest to trigger the first uPy file error might be to
temporarily edit the microfs.py module to manually throw an
Exception in the MicroPython ls() operation. This open an error
pop up, and while that error window is shown, we can disconnect the
USB cable.

What happens here is that because the first error first shows an
error message to the user, it then only tries to close the panel (and
later destroy it) after the user has closed the error message,
giving plenty of time for a different error to happen and
close/destroy the panel.
In normal circumstances these two errors can happen, for example,
if the device itself is busy reboting when the file panel was opened.

Technically this doesn't really fix the race condition, but it
significantly reduces the chances of it happening, as it takes a bit
of time for the panel to be destroyed by the garbage collector, and
this issue normally only occurs when multiple issues are detected
in a short amount of time.