Remove ESP-related USB VIDs & PIDs from other modes.
Unfortunately at the moment there can only be a single mode
associated to a board USB VID & PID for the autotection to work.

This is only for autodection, once a mode is selected, the specific
functionality to run/flash code should still work even if these
values are not in the mode `valid_boards` list.