Fix None being lowercase in checker messages
Consider the following expression:

    foo == None

Before this change, Mu’s checker would tell you “Comparison to none
should be 'if cond is none:'” if it encountered that expression. The
word “none” in that message should be capitalized because it refers to
something that should be capitalized in Python code.

The problem was caused by an inappropriate use of str.capitalize(). The
author probably just wanted to make the first letter of the message
uppercase. Unfortunately, str.capitalize() also makes every other letter
in the string lowercase [1].

[1]: <https://docs.python.org/3.8/library/stdtypes.html#str.capitalize>