Move esptool from user-venv dependency to vendor it with Mu.
Only a single script was needed from the entire esptool package,
and the only dependency needed for that script is pyserial, which
is already included with Mu.
So, vendoring esptool means we don't need to pip download, package,
and install its other dependencies, which have caused trouble in
the past (specially cryptography when it moved to be a rust package).

This changes the "category" of this dependency, from a wheel packaged
with Mu to install in the user venv, to be included as a Mu app runtime
dependency. As esptool is used in the Mu settings, is probably closer
to a "Mu runtime" dependency, than a "user running their code in
a venv" dependency.

Apart from that, esptool and some of its dependencies don't have
wheels in PyPI.
To be able to use the `--platform` flag with pip download it needs
all dependencies to have wheels, so removing this from the user
wheels helps with that as well.