WIP: Prevent Mu from opening more than one instance (#2297)
Implement a Mu-tex (pun definitely intended) by making use of the `QSharedMemory` Qt object which can be used as an interprocess lock.

When Mu starts up, it attempts to acquire that lock. If it can it carries on. If it can't, it logs the fact -- including which other process holds the lock -- and exits.

The user won't usually know that this has even happened as there's no UI element involved. Post-mortem investigation of the logs should show the effect, though.