Implement handshake for running programs (RUN)
The RUN button will reboot the attached device. Depending on the speed
of the device, the speed of the connected computer, the USB tree and
the contents of the boot scripts, this make take some time.
Instead od just sending the script to run after some time, we now first
wait for the python prompt to appear. This brings a robust RUN behavior.