Reset Mutex (Issue 2429) (#2460)
Force the Mu-tex to reset if it's being held by a zombie process

If a process crashes hard before releasing the Mu-tex then a new process won't be able to start. (This is only likely to happen on Linux which doesn't automatically release the shared memory we use to implement the mutex).

This PR always attaches and detaches from the shared memory before checking whether or not it can attach and read the contents. This is enough to take control of the mutex from the zombie process