tests: Fix 'called_once_with' used instead of 'assert_called_once_with'.
A MagicMock object was being asserted with 'called_once_with()'
instead of 'assert_called_once_with()', likely an accidental
mistake.

As a MagicMock always returns another instance of MagicMock on
anything not previoysly set up, the `assert` was always passing,
even though we updated the Window.connect_zoom() implementation
and that should have failed.

In Python 3.12 this type of typo is caught and throws an exception.